+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Deps: update `openai` to 6.25.0.
 - Deps: update `zone.js` to 0.16.1.
 - Deps (dev): update `@types/prismjs` to 1.26.6.
+- Deps (dev): update `karma-jasmine-html-reporter` to 2.2.0.
 
 ### Removed